Widen syscall() parameters as necessary
syscall() needs 64-bit arguments on 64-bit architectures for 64-bit parameters to syscalls.

Strictly speaking we can't even pass NULL there because it can be defined as just "0".
So pass (void*)0 instead.

In practice, on x86-64 this only affects the 6th and 7th parameters to syscall(), because
compilers tend to generate code that zeroes the high 32 bits when passing a 32-bit parameter
in a 64-bit register, and the first five parameters to syscall() are in registers.

Also, on x86-64 gcc tends to generate 'push' instructions that zero-extend the parameters,
hiding the bug, but clang generates 32-bit 'mov' instructions that permit the bug.

+++ src/RecordCommand.cc
@@ -736,7 +736,8 @@ int RecordCommand::run(vector<string>& args) {
     if (flags.nested == NESTED_IGNORE) {
       exec_child(args);
     } else if (flags.nested == NESTED_DETACH) {
-      int ret = syscall(SYS_rrcall_detach_teleport, 0, 0, 0, 0, 0, 0);
+      int ret = syscall(SYS_rrcall_detach_teleport, (uintptr_t)0, (uintptr_t)0,
+        (uintptr_t)0, (uintptr_t)0, (uintptr_t)0, (uintptr_t)0);
       if (ret < 0) {
         FATAL() << "Failed to detach from parent rr";
       }
