+++ src/test/fadvise.c
@@ -7,9 +7,9 @@ int main(void) {
    * since it only provides optimization hints, so this just
    * checks that rr doesn't blow up when it sees one. */
   posix_fadvise(-1, 0, 0, POSIX_FADV_NORMAL);
-  syscall(SYS_fadvise64, -1, 0, 0, POSIX_FADV_NORMAL);
+  syscall(SYS_fadvise64, -1, (size_t)0, (size_t)0, POSIX_FADV_NORMAL);
 #if defined(SYS_fadvise64_64)
-  syscall(SYS_fadvise64_64, -1, POSIX_FADV_NORMAL, 0, 0);
+  syscall(SYS_fadvise64_64, -1, POSIX_FADV_NORMAL, (size_t)0, (size_t)0);
 #endif
 
   atomic_puts("EXIT-SUCCESS");
